feat: create schema file config interface and migrate use fed2 to it
